Code Smell Fix: use explicit casting on integers used for Math.sqrt

+++ src/main/java/io/github/lucasstarsz/fastj/math/Maths.java
@@ -122,7 +122,7 @@ public class Maths {
      * @return The magnitude, as a {@code float} value.
      */
     public static float magnitude(Point p) {
-        return (float) Math.sqrt(p.x * p.x + p.y * p.y);
+        return (float) Math.sqrt((float) (p.x * p.x) + (float) (p.y * p.y));
     }
 
     /**
